requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.24
decorator==4.0.10
dj-database-url==0.4.1
Django==1.9.10
django-filter==0.15.3
djangorestframework==3.5.1
ecdsa==0.13
enum34==1.1.6
Fabric==1.12.0
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
kombu==3.0.37
paramiko==1.17.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
psycopg2==2.6.2
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.1.3
pytz==2016.7
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
whitenoise==3.2.2
